requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
networkx==3.3
shapely==2.0.6
geopandas==1.0.1
requests==2.32.0
numpy==2.1.2
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.9.2
